src/requirements.txt
scikit-learn==1.5.1
matplotlib==3.9.2
numpy==1.26.4
pillow==10.4.0
